refactor: split native syscall handlers by context
